db sync: move SettingsForm to DeviceSettingsForm
* move 'SettingsForm' (UI component) -> 'DeviceSettingsForm'
* move the schema and the store: settings(Schema|Store) -> deviceSettings(Schema|Store)
