+++ apps/web-client/src/lib/forms/index.ts
@@ -1,6 +1,6 @@
 export { default as BookForm } from "./BookForm.svelte";
 export { default as DaisyUIBookForm } from "./DaisyUIBookForm.svelte";
-export { default as SettingsForm } from "./SettingsForm.svelte";
+export { default as DeviceSettingsForm } from "./DeviceSettingsForm.svelte";
 export { default as WarehouseForm } from "./WarehouseForm.svelte";
 export { default as ScannerForm } from "./ScannerForm.svelte";
 export { default as DatabaseDeleteForm } from "./DatabaseDeleteForm.svelte";
