+++ apps/web-client/src/lib/stores/app.ts
@@ -1,5 +1,4 @@
 import { persisted } from "svelte-local-storage-store";
-import { get } from "svelte/store";
 import { BehaviorSubject, catchError, from, map, of, share } from "rxjs";
 import { defaults } from "sveltekit-superforms";
 import { zod } from "sveltekit-superforms/adapters";
@@ -12,18 +11,20 @@ import type { PluginsInterface } from "$lib/plugins";
 import { readableFromStream } from "$lib/utils/streams";
 
 import { LOCAL_STORAGE_APP_SETTINGS, LOCAL_STORAGE_SETTINGS } from "$lib/constants";
-import { settingsSchema } from "$lib/forms/schemas";
+import { deviceSettingsSchema } from "$lib/forms/schemas";
 
 const autoPrintLabelsInner = persisted(LOCAL_STORAGE_APP_SETTINGS, false);
 const toggleAutoprintLabels = () => autoPrintLabelsInner.update((v) => !v);
 export const autoPrintLabels = Object.assign(autoPrintLabelsInner, { toggle: toggleAutoprintLabels });
 
-const { data: defaultSettings } = defaults(zod(settingsSchema));
-export const settingsStore = persisted(LOCAL_STORAGE_SETTINGS, defaultSettings);
+const { data: defaultSettings } = defaults(zod(deviceSettingsSchema));
+export const deviceSettingsStore = persisted<typeof defaultSettings>(LOCAL_STORAGE_SETTINGS, defaultSettings);
 
 const createDBConnectivityStream = () => {
 	const shareSuject = new BehaviorSubject(true);
-	const { couchUrl: url } = get(settingsStore);
+	// TODO: This is commented out -- replace with update to conform to current stack
+	// const { dbUrl: url } = get(deviceSettingsStore);
+	const url = "";
 
 	return browser && url
 		? from(checkUrlConnection(url)).pipe(
