+++ apps/web-client/src/routes/inventory/inbound/[id]/+page.svelte
@@ -34,7 +34,7 @@
 
 	import { type DialogContent, dialogTitle, dialogDescription } from "$lib/dialogs";
 	import { createExtensionAvailabilityStore } from "$lib/stores";
-	import { autoPrintLabels, settingsStore } from "$lib/stores/app";
+	import { autoPrintLabels, deviceSettingsStore } from "$lib/stores/app";
 
 	import { createIntersectionObserver, createTable } from "$lib/actions";
 
@@ -203,10 +203,10 @@
 
 	// #region printing
 	$: handlePrintReceipt = async () => {
-		await printReceipt($settingsStore.receiptPrinterUrl, await getReceiptForNote(db, noteId));
+		await printReceipt($deviceSettingsStore.receiptPrinterUrl, await getReceiptForNote(db, noteId));
 	};
 	$: handlePrintLabel = async (book: BookData) => {
-		await printBookLabel($settingsStore.labelPrinterUrl, book);
+		await printBookLabel($deviceSettingsStore.labelPrinterUrl, book);
 	};
 	// #endregion book-form
