+++ apps/web-client/src/routes/settings/+page.ts
@@ -1,17 +1,14 @@
 import { superValidate } from "sveltekit-superforms/client";
 import { zod } from "sveltekit-superforms/adapters";
+import { get } from "svelte/store";
 
 import type { PageLoad } from "./$types";
-import { get } from "svelte/store";
-import { settingsStore } from "$lib/stores/app";
-import { settingsSchema } from "$lib/forms/schemas";
 
+import { deviceSettingsStore } from "$lib/stores/app";
+import { deviceSettingsSchema } from "$lib/forms/schemas";
 export const load: PageLoad = async () => {
-	const settingsData = get(settingsStore);
+	const deviceSettingsData = get(deviceSettingsStore);
+	const deviceSettingsForm = await superValidate(deviceSettingsData, zod(deviceSettingsSchema));
 
-	const form = await superValidate(settingsData, zod(settingsSchema));
-
-	return {
-		form
-	};
+	return { deviceSettingsForm };
 };
